refactor(internode): delete unused client/registry/service_discovery modules
client.rs, registry.rs, service_discovery.rs, and most of types.rs were
dead code — no callers outside the internode module. ClientError is the
only type still needed (by metaserver_client.rs), so it's moved inline.

Also remove the #[allow(dead_code)] guard on `pub mod internode` and
delete the never-called pinned_allocator/allocator accessor pair.
